codexhorary/backend/requirements-py313.txt
# Core Flask dependencies - Python 3.13 compatible
Flask==2.3.3
Flask-CORS==4.0.0

# Astronomical calculations
pyswisseph==2.10.3.2

# Geographic and timezone support
geopy==2.4.1
# timezonefinder - skipped for Python 3.13 compatibility
pytz==2023.3

# HTTP and utility libraries
requests==2.31.0
python-dateutil==2.8.2

# YAML configuration support
PyYAML==6.0.2

# Production deployment
gunicorn==21.2.0

# Development dependencies
python-dotenv==1.0.0
# watchdog - using latest compatible version
watchdog

# Building executables - using latest available
pyinstaller

# Optional: Enhanced error tracking (uncomment if needed)
# sentry-sdk[flask]==1.32.0

# Optional: Configuration management (uncomment if needed)
# pydantic==2.4.2
# pydantic-settings==2.0.3

# Optional: Enhanced logging (uncomment if needed)
# structlog==23.2.0

# Testing dependencies (uncomment for development)
# pytest==7.4.3
# pytest-flask==1.3.0
# pytest-cov==4.1.0

# Documentation (uncomment if needed)
# sphinx==7.2.6
# sphinx-rtd-theme==1.3.0